requirements_einsteinai4db.txt
# Core dependencies
numpy>=1.19.0
pandas>=1.0.0
scipy>=1.5.0
scikit-learn>=0.23.0
matplotlib>=3.0.0
seaborn>=0.10.0

# Deep learning
# torch>=1.6.0
# torchvision>=0.7.0
# torchaudio>=0.6.0

# Database and data processing
sqlalchemy>=1.3.0
psycopg2-binary>=2.8.0
h5py>=2.10.0
pyarrow>=1.0.0

# Utilities
tqdm>=4.45.0
joblib>=0.15.0
pyyaml>=5.3.1

# Testing
pytest>=5.4.0
pytest-cov>=2.8.0
